Data Objects: Improve IdInventory, IdCodec extensibility
- IdCodec: Improve extensibility
- IdCodec: Handle null/empty string as null-id
- IdCodec: Add abstract test and default test implementation
- IdInventory: Add method to unregister type mapper
- DataObjectSignatureGenerator: Improve extensibility for id resolver
- Add unit tests

322156
